Check for in-progress recovery before backup
Because the etcd backup is being taken from a container, that
container gets restored during recovery and will attempt to take a new
backup when the container is restarted. If a recovery is currently in
progress when the backup is attempted, then, the container can be
safely terminated instead.

Additionally, the recovery utility has been updated to always pause
for a reboot after /etc has been restored, regardless of the /etc
files being updated. This ensures a more consistent and reliable
recovery process, as the cluster restore is then launched from a clean
boot.

Signed-off-by: Don Penney <dpenney@redhat.com>