Remove unneeded recovery utility options
Support for container image backup and restore has been removed from
the recovery utility, as these steps are not required. The required
images would still be available in a local registry, so there's no
need to back them up.

In addition, the option to skip the crio-wipe as part of recovery is
also dropped.

Signed-off-by: Don Penney <dpenney@redhat.com>